// src/interfaces/IEstate.ts
export default interface IEstate {
    name: string;
    size: string;
    oceanDistance: string;
    plotsAvailable: string;
    population: string;
    counties: string;
    paymentTerm: string;
    averagePricePerSQM: string;
    averagePartialPaymentPrice: string;
    id?: number;
    src?: string;
}